IssueVault/requirements.txt
streamlit>=1.35.0
oracledb>=2.2.0
pandas>=2.2.2
plotly>=5.22.0
scikit-learn>=1.5.0
python-dotenv>=1.0.1
